fix(driver): Redesign test driver
In order to react to the situation that transactions' record will disapear,
after snapshot (both global one and local one), the order of test in driver
is redesigned.
Send transaction APIs are ahead of recieve transaction APIs, and the infomation
that are used in the recieve transaction APIs (i.e. tag, bundle hash) are
come from send transaction APIs.