feat(api): Implement find_transaction_objects
In order to take multiple query hashes, we use `iota_client_get_transaction_objects()`
to replace the original get_transaction_object and rename API `get_transaction_objects()` as
`find_transaction_objects()`. This new API receives transaction hases (in a JSON list), and
returns a `transaction_array_t` object which is a list of transaction object can be easily applied.

We use entangled original API `iota_client_find_transactions` to get transaction hashes with
given informations. It uses POST request and request body is the same as IRI API
`findTransactions`, but eliminating the `command` field.

The tag related APIs are deprecated after both `find_trnasctions` and `find_transaction_objects`
are implemeted, so the two deprecated APIs are removed in this commit.