refactor(mam): Refactor api_mam_send_message()
The old implementation of `api_mam_send_message()` is
a wrong implemetaion for our usage.

The current implementaion could automatically send announcement once all the
secret keys under the current given parameters run out.

One thing that we have to improve in the future is the channel depth and endpoint
depth can't be changed once the first channel is created. We should make it changable.