Fix server hang on request to /address
The server hangs when user requests `/address` and IRI connections are
down.

At this moment, we can't shutdown the server by sending `SIGINT` or
`SIGTERM` since served has already registered the signal handler for
those signals.

I added a timeout to api_generate_address, separating a new `/address`
request to a new thread, and simply cancel the thread if exceeded the
time limit.