Implement a table and operate methods for the mechanism of reattach
What :  add a new identity table and corresponding create, insert and select method. A row of the identity table consists of an uuid, a hash ans the TXN status.

Why : In the solution of reattach and permanode, we need a interface to store TXN and could be selected by status and uuid.

How :
@scylla_table.c, @scylla_table.h :
  - define identity table and its' new, free, get/set function.
  - implement UTArray interface to for  identity_t
@scylla_api.c
  - implement create, insert and select function for identity table
  - implement initial function for identity table
@test_scylla.c
  - add test for identity table
d