Add an option for ScyllaDB connection
What :  Connect to the specific ScyllaDB host when performing  initialization. The default setting of connection is disabled.

Use  "--db_host [host]" to specify the host.

The flag "enabled" in db_client_t indicates whether TA successfully connects to the ScyllaDB or not.

Why : To insert new pending transactions and query transactions' status, TA needs connecting to a ScyllaDB host.

Close #348