fix(db) : Revise tests for ScyllaDB backend
- Truncate(clear all data) instead of dropping table. 
- Remove tests for permanode. Those functions do not be used currently.

To test the insertion and selection, we need an empty table. The previous method is dropping old tables when performing tests. But dropping tables will change the schema. After schema changement, we need to wait for the ScyllaDB cluster to make an agreement, but we execute the tests concurrently. It is hard to wait enough time among each test. Actually, the schema should not be changed if we do not change the data model.

close #458