feat(cache): Save failed txns in caching database
Transactions which can't be sent successfully for status code
`SC_CCLIENT_FAILED_RESPONSE` would be cached with redis.

And fecthed transactions are cached with a set object.

Library `uuid-dev` is required after this PR, since we use library
`uuid/uuid.h` to generate an uuid which would be send as response to the
requested client.

close #297