feat(cache): Check cache capacity
`cache_occupied_space()` would return the used size in caching database.

When the capacity of redis server exceeds the maximun capacity, the
health_track thread will pop the oldest UUID and the corresponding
bundle.

To avoid data racing, at the moment that health track is cleaning
redis server, all the other thread is not able to cache service.

Add the Document for buffer service.

close #488