feat(endpoint): Add HMAC support for AES encryption
The HMAC is a specific type of message authentication code(MAC)
involving a cryptographic hash function and a secret cryptographic key.
Same as other MAC, HMAC is used to verify both the data integrity and
the authenticity of a message. We used sha256 as default hash function
to calculate the HMAC.

Two new structure members `hmac` and `timestamp` are added into cipher_ctx.
The timestamp used the source from Unix Epoch. The timestamp is one of
the source to calculate the HMAC.

Close #608