feat(endpoint): Integration with Legato
The files and directory structure are changed.
The directory endpoint/endpointComp is for building the Legato app.
The endpoint/endpoint.c is renamed as endpoint/endpoint_core.c to
separate the core functions. Same as the endpoint/endpoint.h.

New:
To build the endpoint Legato app, use the command 'make legato'.
It is implemented with the mkapp command.

Besides, the shared library libendpoint.so can be generated with Bazel
as well.

Close #576.