fix(Endpoint): Update private key source
This commit changes the hardcoded private key inside the legato targets
and the simulator.

For legato targets, the private will be fetched from secure storage.
The key should be set by the legato shell before the endpoint executes.
If the key doesn't exist in the secure storage, the error would be show.

For the simulator, the private key and the device ID will be generated
into the endpoint.conf if the private key and the device ID doesn't
found.

The secure storage adapter also removed inside this commit, since the
simulator would store the private key and device into the "endpoint.conf"
for simulating the secure storage.

Close #779